requirements.txt


Django==4.0.4
django-allauth==0.50.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-paypal==2.0
django-rest-framework==0.1.0
django-sequences==2.7
djangorestframework==3.13.1
paypal-checkout-serversdk==1.0.1
Pillow==9.1.0
stripe==3.0.0
whitenoise==5.0

